+++ swarm_system/assistants/core/__init__.py
@@ -1,0 +1,8 @@
+"""
+Core assistant implementations for the swarm system.
+"""
+
+from .text_processor import TextProcessorAssistant
+from .calculator_assistant import CalculatorAssistant
+
+__all__ = ["TextProcessorAssistant", "CalculatorAssistant"]
