+++ swarm_system/swarm_demo_report.txt
@@ -1,0 +1,40 @@
+SWARM SYSTEM DEMONSTRATION REPORT
+==================================================
+Generated: 106255.56795
+
+EXECUTION RESULTS:
+✅ Database system operational
+   - Knowledge entries: 1
+   - Cache working: True
+   - Memory storage: ID 1
+✅ Assistant system operational
+   - Registered types: ['text_processor', 'calculator']
+   - Active instances: 2
+   - Text processor result: <class 'str'>
+   - Calculator result: <class 'str'>
+✅ Meta-tooling operational
+   - Dynamic tool creation: 83 chars
+   - Learning storage: 63 chars
+✅ Swarm communication operational
+   - Knowledge retrieval: 201 chars
+   - Cross-component communication: Working
+✅ System status retrieved
+   - Status length: 250 chars
+   - Total database records: 9
+
+SYSTEM CAPABILITIES DEMONSTRATED:
+✅ Database layer (cache.db, memory.db, knowledge.db, coderl.db)
+✅ Assistant registration and management
+✅ Meta-tooling and dynamic tool creation
+✅ Assistant-as-tool functionality
+✅ Knowledge base operations
+✅ Swarm communication protocols
+✅ System monitoring and status
+
+ARCHITECTURE COMPONENTS:
+• Assistant Layer: Base building blocks
+• Agent Layer: Composed assistants (in development)
+• Swarm Layer: Lightweight agent coordination (in development)
+• Meta-Tooling: Dynamic creation capabilities
+• Database Layer: Multi-database management
+• Communication Layer: Inter-component communication
