+++ swarm_system/utils/__init__.py
@@ -1,0 +1,15 @@
+"""
+Utility modules for the swarm system.
+"""
+
+from .database_manager import db_manager
+from .prompts import get_assistant_prompt
+from .tools import (
+    create_dynamic_tool, query_knowledge_base,
+    store_learning, get_swarm_status
+)
+
+__all__ = [
+    "db_manager", "get_assistant_prompt", "create_dynamic_tool",
+    "query_knowledge_base", "store_learning", "get_swarm_status"
+]
