+++ thought_agent_20251001_195637.log
@@ -1,0 +1,3 @@
+2025-10-01 19:56:37 - INFO - Found credentials in shared credentials file: ~/.aws/credentials
+2025-10-01 19:56:37 - INFO - Found credentials in shared credentials file: ~/.aws/credentials
+2025-10-01 19:56:37 - INFO - Found credentials in shared credentials file: ~/.aws/credentials
